feat: Rename checkout button to "Gửi đơn"
Updates the button text from "Thanh toán" to "Gửi đơn" in both the main cart view and the checkout modal. This change clarifies the action, as users are submitting their order to be processed. The checkout button in the modal is also updated to specifically mention "Gửi đơn cho Quân".